pyproject.toml
[project]
name = "soccer-sim"
version = "0.1.0"
description = "A comprehensive soccer simulation system that models the complex dynamics of football leagues, clubs, and player development through sophisticated statistical modeling and dynamic gameplay mechanics."
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
  "async-timeout==4.0.2",
  "cachelib==0.9.0",
  "click==8.1.3",
  "colorama==0.4.5",
  "contourpy==1.0.5",
  "cycler==0.11.0",
  "Deprecated==1.2.13",
  "dnspython==2.2.1",
  "Flask==2.2.2",
  "Flask-Mobility==1.1.0",
  "Flask-Session==0.4.0",
  "fonttools==4.37.3",
  "funcy==1.17",
  "gunicorn==20.1.0",
  "itsdangerous==2.1.2",
  "Jinja2==3.1.2",
  "joblib==1.2.0",
  "kiwisolver==1.4.4",
  "MarkupSafe==2.1.1",
  "matplotlib==3.6.0",
  "numpy==1.23.3",
  "packaging==21.3",
  "Pillow==9.2.0",
  "pymongo==4.2.0",
  "pyparsing==3.0.9",
  "python-dateutil==2.8.2",
  "python-dotenv==0.21.0",
  "redis==4.3.4",
  "rq==1.11.1",
  "six==1.16.0",
  "Werkzeug==2.2.2",
  "wrapt==1.14.1"
]

[dependency-groups]
dev = [
  "ruff==0.9.6",
  "pytest==7.2.0",
  "pre-commit==3.3.3"
]

[tool.ruff]
target-version = "py38"
line-length = 100
lint.select = [
    "E",
    "W",
    "F",
    "I",
    "B",
    "C",
    "UP",
]
lint.ignore = [
    "C901", # function too complex
]
